fix(constant): fix parameter type

+++ model/strategy/params/indicators_params.py
@@ -17,4 +17,4 @@ class CCIParams(BaseModel):
     source_column: str = "close"
     length: int = 20
     ma_type: str = "sma"
-    constant: int = 0.015
+    constant: float = 0.015
